2729 Ajustando a Paginação da Tela de Livros

+++ client/src/pages/Books/index.js
@@ -9,20 +9,23 @@ import logoImage from '../../assets/logo.svg';
 
 export default function Books() {
   const [books, setBooks] = useState([]);
+  const [page, setPage] = useState(1);
 
   const userName = localStorage.getItem('userName');
 
   const accessToken = localStorage.getItem('accessToken');
 
+  const authorization = {
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+    },
+  };
+
   const history = useHistory();
 
   async function logout() {
     try {
-      await api.get(`api/Auth/v1/revoke`, {
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-        },
-      });
+      await api.get(`api/Auth/v1/revoke`, authorization);
 
       localStorage.clear();
       history.push('/');
@@ -41,11 +44,7 @@ export default function Books() {
 
   async function deleteBook(id) {
     try {
-      await api.delete(`api/Book/v1/${id}`, {
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-        },
-      });
+      await api.delete(`api/Book/v1/${id}`, authorization);
 
       setBooks(books.filter((book) => book.id !== id));
     } catch (err) {
@@ -54,14 +53,14 @@ export default function Books() {
   }
 
   useEffect(() => {
-    api
-      .get('api/Book/v1/asc/5/1', {
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-        },
-      })
-      .then((response) => setBooks(response.data.list));
+    fetchMoreBooks();
   }, [accessToken]);
+
+  async function fetchMoreBooks() {
+    const response = await api.get(`api/Book/v1/asc/4/${page}`, authorization);
+    setBooks([...books, ...response.data.list]);
+    setPage(page + 1);
+  }
 
   return (
     <div className="book-container">
@@ -110,6 +109,9 @@ export default function Books() {
           </li>
         ))}
       </ul>
+      <button className="button" onClick={fetchMoreBooks} type="button">
+        Load More
+      </button>
     </div>
   );
 }
